refactor project structuration
