+++ part8/library-gql/server/.eslintrc.json
@@ -10,7 +10,7 @@
   },
   "rules": {
     "indent": ["error", 2],
-    "linebreak-style": ["error", "unix"],
+    "linebreak-style": ["error", "windows"],
     "quotes": ["error", "double"],
     "semi": ["error", "always"]
   }
