+++ part8/library-gql/server/graphql/index.js
@@ -1,0 +1,11 @@
+const { ApolloServer } = require("apollo-server");
+
+const typeDefs = require("./types/index");
+const resolvers = require("./resolvers/index");
+
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+});
+
+module.exports = server;
