+++ part8/library-gql/server/graphql/resolvers/index.js
@@ -1,0 +1,25 @@
+const {
+  allAuthors,
+  authorCount,
+  editAuthor,
+  author,
+} = require("./author.resolver");
+const { allBooks, bookCount, addBook } = require("./book.resolver");
+
+const resolvers = {
+  Query: {
+    bookCount,
+    authorCount,
+    allAuthors,
+    allBooks,
+  },
+  Author: {
+    bookCount: author.bookCount,
+  },
+  Mutation: {
+    addBook,
+    editAuthor,
+  },
+};
+
+module.exports = resolvers;
