+++ part8/library-gql/server/graphql/types/author.types.js
@@ -1,0 +1,12 @@
+const { gql } = require("apollo-server");
+
+const authorTypes = gql`
+  type Author {
+    name: String!
+    born: Int
+    id: ID!
+    bookCount: Int
+  }
+`;
+
+module.exports = authorTypes;
