+++ part8/library-gql/server/graphql/types/book.types.js
@@ -1,0 +1,13 @@
+const { gql } = require("apollo-server");
+
+const bookTypes = gql`
+  type Book {
+    author: Author!
+    genres: [String!]!
+    id: ID!
+    published: Int!
+    title: String!
+  }
+`;
+
+module.exports = bookTypes;
