+++ part8/library-gql/server/graphql/types/mutation.types.js
@@ -1,0 +1,15 @@
+const { gql } = require("apollo-server");
+
+const mutationTypes = gql`
+  type Mutation {
+    addBook(
+      title: String!
+      author: String!
+      published: Int
+      genres: [String!]!
+    ): Book
+    editAuthor(name: String!, born: Int!): Author
+  }
+`;
+
+module.exports = mutationTypes;
