+++ part8/library-gql/server/graphql/types/queries.types.js
@@ -1,0 +1,12 @@
+const { gql } = require("apollo-server");
+
+const queriesTypes = gql`
+  type Query {
+    bookCount: Int!
+    authorCount: Int!
+    allBooks(author: String, genre: String): [Book!]!
+    allAuthors: [Author!]!
+  }
+`;
+
+module.exports = queriesTypes;
