+++ part8/library-gql/server/index.js
@@ -1,184 +1,4 @@
-/* eslint-disable indent */
-const { ApolloServer, gql } = require("apollo-server");
-const { v1: uuid } = require("uuid");
-
-function searchByAuthor(list, author) {
-  return list.filter((item) => item.author === author);
-}
-
-function searchByGenre(list, genre) {
-  return list.filter((item) => item.genres.includes(genre));
-}
-
-function searchByAuthorAndGenre(list, author, genre) {
-  return searchByAuthor(searchByGenre(list, genre), author);
-}
-
-let authors = [
-  {
-    name: "Robert Martin",
-    id: "afa51ab0-344d-11e9-a414-719c6709cf3e",
-    born: 1952,
-  },
-  {
-    name: "Martin Fowler",
-    id: "afa5b6f0-344d-11e9-a414-719c6709cf3e",
-    born: 1963,
-  },
-  {
-    name: "Fyodor Dostoevsky",
-    id: "afa5b6f1-344d-11e9-a414-719c6709cf3e",
-    born: 1821,
-  },
-  {
-    name: "Joshua Kerievsky", // birthyear not known
-    id: "afa5b6f2-344d-11e9-a414-719c6709cf3e",
-  },
-  {
-    name: "Sandi Metz", // birthyear not known
-    id: "afa5b6f3-344d-11e9-a414-719c6709cf3e",
-  },
-];
-
-/*
- * English:
- * It might make more sense to associate a book with its author by storing the author's
- * name in the context of the book instead of the author's id. However, for simplicity,
- * we will store the author's name in connection with the book
- */
-
-let books = [
-  {
-    title: "Clean Code",
-    published: 2008,
-    author: "Robert Martin",
-    id: "afa5b6f4-344d-11e9-a414-719c6709cf3e",
-    genres: ["refactoring"],
-  },
-  {
-    title: "Agile software development",
-    published: 2002,
-    author: "Robert Martin",
-    id: "afa5b6f5-344d-11e9-a414-719c6709cf3e",
-    genres: ["agile", "patterns", "design"],
-  },
-  {
-    title: "Refactoring, edition 2",
-    published: 2018,
-    author: "Martin Fowler",
-    id: "afa5de00-344d-11e9-a414-719c6709cf3e",
-    genres: ["refactoring"],
-  },
-  {
-    title: "Refactoring to patterns",
-    published: 2008,
-    author: "Joshua Kerievsky",
-    id: "afa5de01-344d-11e9-a414-719c6709cf3e",
-    genres: ["refactoring", "patterns"],
-  },
-  {
-    title: "Practical Object-Oriented Design, An Agile Primer Using Ruby",
-    published: 2012,
-    author: "Sandi Metz",
-    id: "afa5de02-344d-11e9-a414-719c6709cf3e",
-    genres: ["refactoring", "design"],
-  },
-  {
-    title: "Crime and punishment",
-    published: 1866,
-    author: "Fyodor Dostoevsky",
-    id: "afa5de03-344d-11e9-a414-719c6709cf3e",
-    genres: ["classic", "crime"],
-  },
-  {
-    title: "The Demon ",
-    published: 1872,
-    author: "Fyodor Dostoevsky",
-    id: "afa5de04-344d-11e9-a414-719c6709cf3e",
-    genres: ["classic", "revolution"],
-  },
-];
-
-const typeDefs = gql`
-  type Author {
-    name: String!
-    born: Int
-    id: ID!
-    bookCount: Int
-  }
-
-  type Book {
-    title: String!
-    author: String!
-    published: Int!
-    genres: [String!]!
-    id: ID!
-  }
-
-  type Query {
-    bookCount: Int!
-    authorCount: Int!
-    allBooks(author: String, genre: String): [Book!]!
-    allAuthors: [Author!]!
-  }
-
-  type Mutation {
-    addBook(
-      title: String!
-      author: String!
-      published: Int
-      genres: [String!]!
-    ): Book
-    editAuthor(name: String!, born: Int!): Author
-  }
-`;
-
-const resolvers = {
-  Query: {
-    bookCount: () => books.length,
-    authorCount: () => authors.length,
-    allAuthors: () => authors,
-    allBooks: (root, args) =>
-      args.author && args.genre
-        ? searchByAuthorAndGenre(books, args.author, args.genre)
-        : args.author
-        ? searchByAuthor(books, args.author)
-        : args.genre
-        ? searchByGenre(books, args.genre)
-        : books,
-  },
-  Author: {
-    bookCount: (root) =>
-      books.reduce(
-        (accumulator, book) =>
-          book.author === root.name ? accumulator + 1 : accumulator,
-        0
-      ),
-  },
-  Mutation: {
-    addBook: (root, args) => {
-      if (!authors.map((author) => author.name).includes(args.author))
-        authors = authors.concat({ name: args.author, id: uuid() });
-      const book = { ...args, id: uuid() };
-      books = books.concat(book);
-      return book;
-    },
-    editAuthor: (root, args) => {
-      if (!authors.find((author) => author.name === args.name)) return null;
-      const foundAuthor = authors.find((author) => author.name === args.name);
-      const newAuthor = { ...foundAuthor, ...args };
-      authors = authors.map((author) =>
-        author.name === args.name ? newAuthor : author
-      );
-      return newAuthor;
-    },
-  },
-};
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-});
+const server = require("./graphql/index");
 
 server.listen().then(({ url }) => {
   console.log(`Server ready at ${url}`);
